+++ examples/next-ts/pages/color-picker.tsx
@@ -31,6 +31,7 @@ export default function Page() {
     colorPicker.machine({
       id: useId(),
       name: "color",
+      format: "hsla",
       value: colorPicker.parse("hsl(0, 100%, 50%)"),
     }),
     {