+++ packages/machines/toast/src/toast-group.connect.ts
@@ -3,16 +3,21 @@ import type { NormalizeProps, PropTypes } from "@zag-js/types"
 import { runIfFn, uuid } from "@zag-js/utils"
 import { parts } from "./toast.anatomy"
 import { dom } from "./toast.dom"
-import type { GroupMachineApi, GroupSend, GroupState, Options } from "./toast.types"
+import type {
+  DefaultGenericOptions,
+  GenericOptions,
+  GroupMachineApi,
+  GroupSend,
+  GroupState,
+  Options,
+} from "./toast.types"
 import { getGroupPlacementStyle, getToastsByPlacement } from "./toast.utils"
 
-export let toaster = {} as GroupMachineApi
-
-export function groupConnect<T extends PropTypes>(
-  state: GroupState,
+export function groupConnect<T extends PropTypes, O extends GenericOptions = DefaultGenericOptions>(
+  state: GroupState<O>,
   send: GroupSend,
   normalize: NormalizeProps<T>,
-): GroupMachineApi<T> {
+): GroupMachineApi<T, O> {
   //
   const toastsByPlacement = getToastsByPlacement(state.context.toasts)
 
@@ -21,7 +26,7 @@ export function groupConnect<T extends PropTypes>(
     return !!state.context.toasts.find((toast) => toast.id == id)
   }
 
-  function create(options: Options) {
+  function create(options: Options<O>) {
     const uid = `toast:${uuid()}`
     const id = options.id ? options.id : uid
 
@@ -31,13 +36,13 @@ export function groupConnect<T extends PropTypes>(
     return id
   }
 
-  function update(id: string, options: Options) {
+  function update(id: string, options: Options<O>) {
     if (!isVisible(id)) return
     send({ type: "UPDATE_TOAST", id, toast: options })
     return id
   }
 
-  function upsert(options: Options) {
+  function upsert(options: Options<O>) {
     const { id } = options
     const visible = id ? isVisible(id) : false
     if (visible && id != null) {