+++ website/next.config.js
@@ -4,7 +4,6 @@ const { withContentlayer } = require("next-contentlayer")
  * @type {import('next').NextConfig}
  */
 const nextConfig = {
-  swcMinify: false,
   experimental: {
     externalDir: true,
   },