+++ packages/machines/hover-card/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @zag-js/hover-card
 
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@0.30.0
+  - @zag-js/core@0.30.0
+  - @zag-js/types@0.30.0
+  - @zag-js/utils@0.30.0
+  - @zag-js/dismissable@0.30.0
+  - @zag-js/dom-query@0.30.0
+  - @zag-js/popper@0.30.0
+
 ## 0.29.0
 
 ### Minor Changes