+++ packages/utilities/focus-visible/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zag-js/focus-visible
 
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/dom-query@0.30.0
+
 ## 0.29.0
 
 ### Patch Changes