+++ packages/utilities/remove-scroll/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zag-js/remove-scroll
 
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/dom-query@0.30.0
+
 ## 0.29.0
 
 ### Patch Changes